Add defaults and validators for all valid config keys
This adds a lazily-loaded dictionary containing all valid fedmsg
configuration keys along with validators for all those keys to ensure
they are set to an acceptable value.

This new configuration dictionary is meant to replace the old
load_config API. That API has been marked as deprecated and will be
removed in a future release.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>